testing/mochitest/tests/Harness_sanity/test_sanityRegisteredServiceWorker2.html

<!DOCTYPE html>
<html>
<head>
  <title>Test that service worker registrations not cleaned up in mochitests cause failures</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<script>
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [
  ["dom.serviceWorkers.exemptFromPerDomainMax", true],
  ["dom.serviceWorkers.interception.enabled", true],
  ["dom.serviceWorkers.enabled", true],
  ["dom.serviceWorkers.testing.enabled", true]
]}, function() {
  navigator.serviceWorker.getRegistration("scope")
    .then(function(registration) {
      ok(registration, "Registration successfully obtained");
      return registration.unregister();
    }).then(function() {
      SimpleTest.finish();
    });
});
</script>
</body>
</html>
